feat(parameter): Add parameter management utilities for robot identification
